Move buffer-related logic from worktree to project
